requirements.txt
streamlit==1.32.2
python-dotenv==1.0.1
python-docx==1.1.0
PyPDF2==3.0.1
google-genai==0.3.0